Different method for identifying non-detect quantiles
The old method didn't work when handling chemicals with variable LODs
(e.g. any lipid adjusted readings.)

This method should work in those cases, but there appear to be
discrepancies between it and the summary tables (specifically I was
checking against PCB 153.) I'm waiting to address this until I get more
of a sense of how big the problem is, which I'll see when I can compare
everything to the summary tables programmatically.